refactor to own ConfigController and ConfigService
