+++ src/main/java/com/moorhuhnservice/moorhuhnservice/data/ConfigurationDTO.java
@@ -1,6 +1,5 @@
 package com.moorhuhnservice.moorhuhnservice.data;
 
-import java.util.List;
 import java.util.Set;
 import lombok.AllArgsConstructor;
 import lombok.Data;
