+++ src/main/java/com/moorhuhnservice/moorhuhnservice/data/mapper/ConfigurationMapper.java
@@ -2,7 +2,6 @@ package com.moorhuhnservice.moorhuhnservice.data.mapper;
 
 import com.moorhuhnservice.moorhuhnservice.data.Configuration;
 import com.moorhuhnservice.moorhuhnservice.data.ConfigurationDTO;
-import java.util.List;
 import org.mapstruct.Mapper;
 
 @Mapper(componentModel = "spring")
