+++ src/main/java/com/moorhuhnservice/moorhuhnservice/data/mapper/QuestionMapper.java
@@ -2,7 +2,6 @@ package com.moorhuhnservice.moorhuhnservice.data.mapper;
 
 import com.moorhuhnservice.moorhuhnservice.data.Question;
 import com.moorhuhnservice.moorhuhnservice.data.QuestionDTO;
-import java.util.List;
 import java.util.Set;
 import org.mapstruct.Mapper;
