+++ src/main/java/com/moorhuhnservice/moorhuhnservice/repositories/ConfigurationRepository.java
@@ -1,8 +1,6 @@
 package com.moorhuhnservice.moorhuhnservice.repositories;
 
 import com.moorhuhnservice.moorhuhnservice.data.Configuration;
-import com.moorhuhnservice.moorhuhnservice.data.Question;
-import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
